Spliet the Binding Form Data of the Exception Dialog into a seperate file. Consitant with other places in the application.
